Move Module\Profile to Module\Profile\Index
- Move /profiles to Module\Settings\Profile\Index
