+++ mod/settings.php
@@ -74,9 +74,9 @@ function settings_init(App $a)
 	];
 
 	$tabs[] =	[
-		'label'	=> DI::l10n()->t('Profiles'),
-		'url' 	=> 'profiles',
-		'selected'	=> (($a->argc == 1) && ($a->argv[0] === 'profiles')?'active':''),
+		'label'	=> DI::l10n()->t('Profile'),
+		'url' 	=> 'settings/profile',
+		'selected'	=> (($a->argc > 1) && ($a->argv[1] === 'profile')?'active':''),
 		'accesskey' => 'p',
 	];
