--- view/templates/profile-hide-wall.tpl
@@ -1,17 +1,0 @@
-
-<p id="hide-wall-text">
-{{$desc nofilter}}
-</p>
-
-		<div id="hide-wall-yes-wrapper">
-		<label id="hide-wall-yes-label" for="hide-wall-yes">{{$yes_str}}</label>
-		<input type="radio" name="hidewall" id="hide-wall-yes" {{$yes_selected}} value="1" />
-
-		<div id="hide-wall-break" ></div>	
-		</div>
-		<div id="hide-wall-no-wrapper">
-		<label id="hide-wall-no-label" for="hide-wall-no">{{$no_str}}</label>
-		<input type="radio" name="hidewall" id="hide-wall-no" {{$no_selected}} value="0"  />
-
-		<div id="hide-wall-end"></div>
-		</div>
