--- view/templates/profile-in-directory.tpl
@@ -1,17 +1,0 @@
-
-<p id="profile-in-directory">
-{{$desc nofilter}}
-</p>
-
-		<div id="profile-in-dir-yes-wrapper">
-		<label id="profile-in-dir-yes-label" for="profile-in-dir-yes">{{$yes_str}}</label>
-		<input type="radio" name="profile_in_directory" id="profile-in-dir-yes" {{$yes_selected}} value="1" />
-
-		<div id="profile-in-dir-break" ></div>	
-		</div>
-		<div id="profile-in-dir-no-wrapper">
-		<label id="profile-in-dir-no-label" for="profile-in-dir-no">{{$no_str}}</label>
-		<input type="radio" name="profile_in_directory" id="profile-in-dir-no" {{$no_selected}} value="0"  />
-
-		<div id="profile-in-dir-end"></div>
-		</div>
