--- view/templates/profile-in-netdir.tpl
@@ -1,17 +1,0 @@
-
-<p id="profile-in-directory">
-{{$desc nofilter}}
-</p>
-
-		<div id="profile-in-netdir-yes-wrapper">
-		<label id="profile-in-netdir-yes-label" for="profile-in-netdir-yes">{{$yes_str}}</label>
-		<input type="radio" name="profile_in_netdirectory" id="profile-in-netdir-yes" {{$yes_selected}} value="1" />
-
-		<div id="profile-in-netdir-break" ></div>	
-		</div>
-		<div id="profile-in-netdir-no-wrapper">
-		<label id="profile-in-netdir-no-label" for="profile-in-netdir-no">{{$no_str}}</label>
-		<input type="radio" name="profile_in_netdirectory" id="profile-in-netdir-no" {{$no_selected}} value="0"  />
-
-		<div id="profile-in-netdir-end"></div>
-		</div>
