--- view/templates/profile_edlink.tpl
@@ -1,3 +1,0 @@
-
-<div class="profile-edit-side-div"><a class="profile-edit-side-link icon edit" title="{{$editprofile}}" href="profiles/{{$profid}}" ></a></div>
-<div class="clear"></div>
