+++ view/templates/settings/profile/hide-friends.tpl
