+++ view/templates/settings/profile/link.tpl
@@ -1,0 +1,2 @@
+<div class="profile-edit-side-div"><a class="profile-edit-side-link icon edit" title="{{$editprofile}}" href="settings/profile"></a></div>
+<div class="clear"></div>
