+++ view/templates/welcome.tpl
@@ -26,11 +26,11 @@
 				{{$profile_photo_txt nofilter}}
 			</li>
 			<li>
-				<a target="newmember" href="profiles">{{$profiles_link}}</a><br />
+				<a target="newmember" href="settings/profile">{{$profiles_link}}</a><br />
 				{{$profiles_txt nofilter}}
 			</li>
 			<li>
-				<a target="newmember" href="profiles">{{$profiles_keywords_link}}</a><br />
+				<a target="newmember" href="settings/profile">{{$profiles_keywords_link}}</a><br />
 				{{$profiles_keywords_txt nofilter}}
 			</li>
 		</ul>
